add new torch arch
